Add typedef
In the following example, x is defined as an alias for int.

typedef x;

Below is valid C code where the second t is a local variable
of type int having value 3.

typedef int t;
t t = 3;

+++ include/core.h
@@ -88,7 +88,7 @@ class Token {
 
   Token *skip(char *s);
 
-  int get_number();
+  long get_number();
 
   bool consume(Token **rest, char *str);
 };
